@@ -3,38 +3,26 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
-    "@fortawesome/react-fontawesome": "^0.1.14",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@types/enzyme": "^3.10.8",
-    "@types/jest": "^26.0.20",
-    "@types/jwt-decode": "^2.2.1",
-    "@types/lodash": "^4.14.168",
-    "@types/node": "^14.14.28",
-    "@types/react": "^16.14.4",
-    "@types/react-animate-on-scroll": "^2.1.3",
-    "@types/react-dom": "^16.9.11",
-    "@types/react-router-dom": "^5.1.7",
-    "chai": "^4.3.0",
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@types/jwt-decode": "^3.1.0",
+    "@types/node": "^18.0.3",
+    "@types/react": "^18.0.15",
+    "@types/react-animate-on-scroll": "^2.1.5",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-router-dom": "^5.3.3",
     "classnames": "^2.3.1",
-    "dotenv": "^8.2.0",
-    "jwt-decode": "^2.2.0",
-    "lodash": "^4.17.21",
-    "mocha": "^7.1.2",
-    "node-forge": "^1.3.0",
-    "react": "^17.0.1",
-    "react-animate-on-scroll": "^2.1.5",
-    "react-dom": "^17.0.1",
-    "react-error-boundary": "^3.1.3",
-    "react-hot-toast": "^2.0.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.2",
+    "dotenv": "^16.0.1",
+    "jwt-decode": "^3.1.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-error-boundary": "^3.1.4",
+    "react-hot-toast": "^2.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
     "sass": "^1.53.0",
-    "supertest": "^4.0.2",
-    "typescript": "^3.9.9"
+    "typescript": "^4.7.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -60,12 +48,24 @@
   },
   "devDependencies": {
     "@babel/helper-regex": "^7.10.5",
-    "@typescript-eslint/eslint-plugin": "^4.28.3",
-    "@typescript-eslint/parser": "^4.28.3",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.1",
+    "@types/enzyme": "^3.10.12",
+    "@types/jest": "^28.1.4",
+    "@typescript-eslint/eslint-plugin": "^5.30.5",
+    "@typescript-eslint/parser": "^5.30.5",
+    "chai": "^4.3.6",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.6",
     "enzyme-to-json": "^3.6.1",
-    "eslint": "^7.30.0",
-    "eslint-plugin-jsx-a11y": "^6.4.1"
+    "eslint": "^8.19.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "mocha": "^10.0.0",
+    "prettier": "^2.7.1",
+    "supertest": "^6.2.4"
   }
 }
