@@ -1,39 +1,16 @@
-import React, { LazyExoticComponent, FunctionComponent } from 'react';
-import { Route, Redirect, RouteComponentProps } from 'react-router-dom';
-import { StaticContext } from 'react-router';
+import React, { FunctionComponent } from 'react';
+import { Navigate } from 'react-router-dom';
 import TokenService from '../../services/token-service';
 
-export interface PublicOnlyRouteProps {
-  /** JSX Component */
-  component: LazyExoticComponent<
-    FunctionComponent<
-      RouteComponentProps<
-        { [x: string]: string | undefined },
-        StaticContext,
-        any
-      >
-    >
-  >;
-  path: string;
-}
+type PublicOnlyRouteProps = {
+  /** Passed in Component to display */
+  Component: FunctionComponent;
+};
 
 const PublicOnlyRoute: FunctionComponent<PublicOnlyRouteProps> = ({
-  component,
-  ...props
+  Component,
 }): JSX.Element => {
-  const Component = component;
-  return (
-    <Route
-      {...props}
-      render={(componentProps) =>
-        TokenService.hasAuthToken() ? (
-          <Redirect to={'/'} />
-        ) : (
-          <Component {...componentProps} />
-        )
-      }
-    />
-  );
+  return TokenService.hasAuthToken() ? <Navigate to='/' /> : <Component />;
 };
 
 export default PublicOnlyRoute;
