@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from 'react';
 import { Link } from 'react-router-dom';
-import Image from '../Image';
 import PokeTeamsLogo from '../../Images/PokeTeams.png';
+import Image from '../Image';
 import styles from './Footer.module.scss';
 
 const Footer: FunctionComponent = () => {
@@ -10,7 +10,7 @@ const Footer: FunctionComponent = () => {
       <Image
         imageClass={styles['small-logo-image']}
         src={PokeTeamsLogo}
-        alt="PokeTeams Logo"
+        alt='PokeTeams Logo'
       />
       <span>{`Created by Daniel Sottile || © August 2021 || Pokémon is © 1995-2021 Nintendo`}</span>
       <div>
