@@ -1,7 +1,10 @@
-import React, { FunctionComponent, ReactNode, ReactPortal } from 'react';
-import { createPortal } from 'react-dom';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import React, {
+  FunctionComponent,
+  ReactNode /** @todo: ReactPortal **/,
+} from 'react';
 import { faWindowClose } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { createPortal } from 'react-dom';
 import Button from '../Button';
 import styles from './Modal.module.scss';
 
@@ -17,7 +20,9 @@ const Modal: FunctionComponent<ModalProps> = ({
   children,
   isModalOpen,
   handleSetModal,
-}): ReactPortal | null => {
+  /** @todo: The types of ReactPortal and the types in react-dom are mismatched */
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+}): any => {
   if (!isModalOpen) {
     return null;
   }
