@@ -33,9 +33,9 @@ const PokemonValueInputs: FunctionComponent<PokemonValueInputsProps> = ({
           onChangeCallback={value.onChangeCallback}
           placeholder={`${value.isEffortValue ? '0' : '31'}`}
           value={value.value}
-          type="text"
+          type='text'
           name={value.name}
-          min="0"
+          min='0'
           max={`${value.isEffortValue ? '252' : '31'}`}
           id={value.id}
           disabled={isPublic}
