@@ -1,17 +1,17 @@
 import React, { useContext, FunctionComponent, MouseEvent } from 'react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
   faSearch,
   faSort,
   faExclamation,
 } from '@fortawesome/free-solid-svg-icons';
-import GeneralContext from '../../contexts/GeneralContext';
-import Input from '../Input';
-import Button from '../Button';
-import Select from '../Select';
-import { validateSearch } from '../../utils/validations';
-import styles from './SearchBar.module.scss';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { TextInput } from '../../@types';
+import GeneralContext from '../../contexts/GeneralContext';
+import { validateSearch } from '../../utils/validations';
+import Button from '../Button';
+import Input from '../Input';
+import Select from '../Select';
+import styles from './SearchBar.module.scss';
 
 type SearchBarProps = {
   /** passed in from Navigation, determines whether to use search or filter */
@@ -60,23 +60,23 @@ const SearchBar: FunctionComponent<SearchBarProps> = ({ isPublic }) => {
             labelClass={styles['label-font']}
             label={isPublic ? 'Search: ' : 'Filter: '}
             labelIcon={<FontAwesomeIcon icon={faSearch} />}
-            placeholder="e.g. Pikachu"
+            placeholder='e.g. Pikachu'
             value={searchInput.value}
             onChangeCallback={(e) =>
               setSearchOrFilter({ value: e.target.value, touched: true })
             }
-            type="text"
-            name="search"
-            id="search"
+            type='text'
+            name='search'
+            id='search'
           />
           <Select
             selectHasError={false}
             containerClass={styles['select']}
-            htmlFor="sort"
-            label="Sort By: "
+            htmlFor='sort'
+            label='Sort By: '
             labelIcon={<FontAwesomeIcon icon={faSort} />}
-            name="sort"
-            id="sort"
+            name='sort'
+            id='sort'
             value={sortOrFilterSort.value}
             onChangeCallback={(e) =>
               setSortOrFilterSort({ value: e.target.value, touched: true })
@@ -90,8 +90,8 @@ const SearchBar: FunctionComponent<SearchBarProps> = ({ isPublic }) => {
           />
           <Button
             buttonClass={styles['go-button']}
-            id="go-button"
-            type="submit"
+            id='go-button'
+            type='submit'
             onClickCallback={(e) => {
               handleSearchOrFilter(e);
             }}
