@@ -1,35 +1,32 @@
 import React, { FunctionComponent } from 'react';
-import { StaticContext } from 'react-router';
-import { Link, RouteComponentProps } from 'react-router-dom';
-import { Location, History, LocationDescriptorObject } from 'history';
-import ScrollAnimation from 'react-animate-on-scroll';
-import Image from '../../components/Image/Image';
+import { Link, useLocation, useNavigate } from 'react-router-dom';
 import Footer from '../../components/Footer/Footer';
+import Image from '../../components/Image/Image';
 import LoginForm from '../../components/LoginForm/LoginForm';
+import AnimateOnScroll from '../../components/Utils/AnimateOnScroll';
 import styles from './LandingPage.module.scss';
 
-type RouteState = {
-  location: Location;
-  history: History;
-  from: LocationDescriptorObject<RouteState>;
-};
+interface LocationState {
+  from: {
+    pathname: string;
+  };
+}
 
-const LandingPage: FunctionComponent<
-  RouteComponentProps<
-    { [x: string]: string | undefined },
-    StaticContext,
-    RouteState
-  >
-> = ({ location, history }): JSX.Element => {
+const LandingPage: FunctionComponent = (): JSX.Element => {
+  const location = useLocation();
+  const navigate = useNavigate();
+
   const handleLoginSucess = (): void => {
-    const destination: LocationDescriptorObject<RouteState> =
-      (location.state || {}).from || '/';
-    history.push(destination);
+    const { from } = (location.state as LocationState) || {
+      from: { pathname: '/' },
+    };
+
+    navigate(from);
   };
 
   return (
     <>
-      <header role="banner" className={styles['header']}>
+      <header role='banner' className={styles['header']}>
         <div className={styles['landing-title']}></div>
         <h2 className={styles['header-text']}>store. organize. edit. share.</h2>
       </header>
@@ -53,13 +50,13 @@ const LandingPage: FunctionComponent<
           <LoginForm onLoginSuccess={handleLoginSucess} />
           <div>
             <p>Not A Member? </p>
-            <Link className={styles['register-entry']} to="/register">
+            <Link className={styles['register-entry']} to='/register'>
               Register Here:
             </Link>
           </div>
           <div>
             <p>Or...</p>
-            <Link className={styles['public-entry']} to="/">
+            <Link className={styles['public-entry']} to='/'>
               Find Teams Here!
             </Link>
           </div>
@@ -67,14 +64,14 @@ const LandingPage: FunctionComponent<
       </section>
 
       <section>
-        <ScrollAnimation animateIn={styles['fade-in']} animateOnce={true}>
+        <AnimateOnScroll animateIn={styles['fade-in']} animateOnce={true}>
           <header>
             <h3>Store your teams like never before!</h3>
           </header>
           <Image
             imageClass={styles['landingimg']}
-            src="https://imgur.com/9wYBaMQ.png"
-            alt="organize page"
+            src='https://imgur.com/9wYBaMQ.png'
+            alt='organize page'
           />
           <p>
             PokeTeams goes above and beyond the average pastebin in order for
@@ -82,60 +79,60 @@ const LandingPage: FunctionComponent<
             into a secure database. Never lose a team again! Say goodbye to
             Pastebin's forever!
           </p>
-        </ScrollAnimation>
+        </AnimateOnScroll>
       </section>
 
       <section>
-        <ScrollAnimation animateIn={styles['fade-in']} animateOnce={true}>
+        <AnimateOnScroll animateIn={styles['fade-in']} animateOnce={true}>
           <header>
             <h3>Upload, Edit, and Maintain your teams!</h3>
           </header>
           <Image
             imageClass={styles['landingimg']}
-            src="https://imgur.com/HR1yjKA.png"
-            alt="edit set"
+            src='https://imgur.com/HR1yjKA.png'
+            alt='edit set'
           />
           <p>
             Once you've uploaded your teams, you'll be able to edit them by hand
             or by importing individual sets! The possibilities are endless!
           </p>
-        </ScrollAnimation>
+        </AnimateOnScroll>
       </section>
 
       <section>
-        <ScrollAnimation animateIn={styles['fade-in']} animateOnce={true}>
+        <AnimateOnScroll animateIn={styles['fade-in']} animateOnce={true}>
           <header>
             <h3>Search By Numerous Parameters!</h3>
           </header>
           <Image
             imageClass={styles['landingimg']}
-            src="https://imgur.com/JwMxPdp.png"
-            alt="search"
+            src='https://imgur.com/JwMxPdp.png'
+            alt='search'
           />
           <p>
             PokeTeams lets you search the database by specific Pokemon species,
             alphabetical order and date created. In the future, you'll be able
             to search for specific teams and by popularity!
           </p>
-        </ScrollAnimation>
+        </AnimateOnScroll>
       </section>
 
       <section>
-        <ScrollAnimation animateIn={styles['fade-in']} animateOnce={true}>
+        <AnimateOnScroll animateIn={styles['fade-in']} animateOnce={true}>
           <header>
             <h3>Share Your Teams and Pokemon!</h3>
           </header>
           <Image
             imageClass={styles['landingimg']}
-            src="https://imgur.com/JWIgCyL.png"
-            alt="export"
+            src='https://imgur.com/JWIgCyL.png'
+            alt='export'
           />
           <p>
             Most importantly, you can share both full teams and Pokemon sets
             with anyone! Create awesome teams and Pokemon and let the world know
             about your creations!
           </p>
-        </ScrollAnimation>
+        </AnimateOnScroll>
       </section>
 
       <Footer />
