@@ -1,24 +1,20 @@
 import React, { useState, useEffect, FunctionComponent } from 'react';
-import { Link, RouteComponentProps } from 'react-router-dom';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faHome } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { Link, useParams } from 'react-router-dom';
+import { PokemonFolder, PokemonTeam, PokemonSet } from '../../@types';
 import FolderPublicShare from '../../components/FolderPublicShare';
-import * as array from 'lodash';
 import apiService from '../../services/apiService';
 import styles from './ShareFolderPage.module.scss';
-import { PokemonFolder, PokemonTeam, PokemonSet } from '../../@types';
 
-export type MatchParams = { folder_id: string };
-
-const ShareFolderPage: FunctionComponent<RouteComponentProps<MatchParams>> = ({
-  match,
-}): JSX.Element => {
+const ShareFolderPage: FunctionComponent = (): JSX.Element => {
+  const { folder_id } = useParams();
   const [folder, setFolder] = useState<PokemonFolder[]>([]);
   const [teams, setTeams] = useState<PokemonTeam[]>([]);
   const [sets, setSets] = useState<PokemonSet[]>([]);
 
   useEffect(() => {
-    const id = match.params.folder_id;
+    const id = folder_id;
     apiService
       .getSingleFolderPublic(Number(id))
       .then((data) => {
@@ -39,12 +35,12 @@ const ShareFolderPage: FunctionComponent<RouteComponentProps<MatchParams>> = ({
             });
 
             Promise.all(promiseArray).then((values: PokemonSet[][]) => {
-              const merged: PokemonSet[] = array.flattenDeep(values);
+              const merged: PokemonSet[] = values.flat();
               setSets(merged);
             });
           });
       });
-  }, [match.params.folder_id]);
+  }, [folder_id]);
 
   return (
     <div className={styles['container']}>
