@@ -1,27 +1,21 @@
 import React, { FunctionComponent, useState, useEffect } from 'react';
-import { Link, RouteComponentProps } from 'react-router-dom';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faHome } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { Link, useParams } from 'react-router-dom';
+import { PokemonSet } from '../../@types';
 import PokemonSetForm from '../../components/PokemonSetForm';
 import apiService from '../../services/apiService';
 import styles from './ShareSetPage.module.scss';
-import { PokemonSet } from '../../@types';
 
-export type MatchParams = {
-  /** match param */
-  set_id: string;
-};
-
-const ShareSetPage: FunctionComponent<RouteComponentProps<MatchParams>> = ({
-  match,
-}): JSX.Element => {
+const ShareSetPage: FunctionComponent = (): JSX.Element => {
+  const { set_id } = useParams();
   const [set, setPokemonSet] = useState<PokemonSet | null>(null);
 
   useEffect(() => {
-    apiService.getSingleSet(Number(match.params.set_id)).then((data) => {
+    apiService.getSingleSet(Number(set_id)).then((data) => {
       setPokemonSet(data);
     });
-  }, [match.params.set_id]);
+  }, [set_id]);
 
   return (
     <div className={styles['container']}>
