@@ -1,5 +1,5 @@
 import { useRef } from 'react';
-import toast from 'react-hot-toast';
+import { toast } from 'react-hot-toast';
 
 interface ClipboardReturn {
   copySuccess: () => string;
